requirements.txt
aiohttp
jinja2
docker-py
